+++ enterprise/api_client/enterprise_catalog.py
@@ -6,6 +6,7 @@ Client for communicating with the Enterprise API.
 from __future__ import absolute_import, unicode_literals
 
 import json
+from collections import OrderedDict
 from logging import getLogger
 
 from requests.exceptions import ConnectionError, Timeout  # pylint: disable=redefined-builtin
@@ -13,6 +14,7 @@ from slumber.exceptions import HttpNotFoundError, SlumberBaseException
 
 from django.conf import settings
 
+from enterprise import utils
 from enterprise.api_client.lms import JwtLmsApiClient
 
 LOGGER = getLogger(__name__)
@@ -25,7 +27,13 @@ class EnterpriseCatalogApiClient(JwtLmsApiClient):
 
     API_BASE_URL = settings.ENTERPRISE_CATALOG_INTERNAL_ROOT_URL + '/api/v1/'
     ENTERPRISE_CATALOG_ENDPOINT = 'enterprise-catalogs'
+    GET_CONTENT_METADATA_ENDPOINT = ENTERPRISE_CATALOG_ENDPOINT + '/{}/get_content_metadata'
+    ENTERPRISE_CUSTOMER_ENDPOINT = 'enterprise-customer'
     APPEND_SLASH = True
+
+    def __init__(self, user=None):
+        user = user if user else utils.get_enterprise_worker_user()
+        super(EnterpriseCatalogApiClient, self).__init__(user)
 
     @JwtLmsApiClient.refresh_token
     def create_enterprise_catalog(
@@ -99,3 +107,58 @@ class EnterpriseCatalogApiClient(JwtLmsApiClient):
                 catalog_uuid, str(exc)
             )
             return {}
+
+    @JwtLmsApiClient.refresh_token
+    def get_content_metadata(self, enterprise_customer, enterprise_catalogs=None):
+        """
+        Return all content metadata contained in the catalogs associated with the EnterpriseCustomer.
+
+        Arguments:
+            enterprise_customer (EnterpriseCustomer): The EnterpriseCustomer to return content metadata for.
+            enterprise_catalogs (EnterpriseCustomerCatalog): Optional list of EnterpriseCustomerCatalog objects.
+
+        Returns:
+            list: List of dicts containing content metadata.
+        """
+        content_metadata = OrderedDict()
+        enterprise_customer_catalogs = enterprise_catalogs or enterprise_customer.enterprise_customer_catalogs.all()
+
+        for enterprise_customer_catalog in enterprise_customer_catalogs:
+            catalog_uuid = enterprise_customer_catalog.uuid
+            endpoint = getattr(self.client, self.GET_CONTENT_METADATA_ENDPOINT.format(catalog_uuid))
+            try:
+                response = endpoint.get()
+                for item in response['results']:
+                    content_id = utils.get_content_metadata_item_id(item)
+                    content_metadata[content_id] = item
+            except (SlumberBaseException, ConnectionError, Timeout) as exc:
+                LOGGER.exception(
+                    'Failed to get content metadata for Catalog [%s] in enterprise-catalog due to: [%s]',
+                    catalog_uuid, str(exc)
+                )
+
+        return list(content_metadata.values())
+
+    @JwtLmsApiClient.refresh_token
+    def contains_content_items(self, catalog_uuid, content_ids):
+        """
+        Checks whether an enterprise catalog contains the given content
+
+        The enterprise catalog endpoint does not differentiate between course_run_ids and program_uuids so they can
+        be used interchangeably. The two query parameters are left in for backwards compatability with edx-enterprise.
+        """
+        query_params = {'course_run_ids': content_ids}
+        endpoint = getattr(self.client, self.ENTERPRISE_CATALOG_ENDPOINT)(catalog_uuid)
+        return endpoint.contains_content_items.get(**query_params)['contains_content_items']
+
+    @JwtLmsApiClient.refresh_token
+    def enterprise_contains_content_items(self, enterprise_uuid, content_ids):
+        """
+        Checks whether an enterprise customer has any catalogs that contain the provided content ids.
+
+        The endpoint does not differentiate between course_run_ids and program_uuids so they can be used
+        interchangeably. The two query parameters are left in for backwards compatability with edx-enterprise.
+        """
+        query_params = {'course_run_ids': content_ids}
+        endpoint = getattr(self.client, self.ENTERPRISE_CUSTOMER_ENDPOINT)(enterprise_uuid)
+        return endpoint.contains_content_items.get(**query_params)['contains_content_items']